requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
idna==2.10
lxml==4.6.2
numpy==1.20.1
pandas==1.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
slack-sdk==3.2.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
